requirements.txt
django
dependency_injector
djangorestframework
django-filter
amqpstorm